feat: resolve comments v2
