+++ gorilla-minitwit/go.mod
@@ -5,7 +5,6 @@ go 1.23.1
 require (
 	github.com/gorilla/mux v1.8.1
 	github.com/gorilla/sessions v1.2.0
-	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 	golang.org/x/crypto v0.17.0
 	gorm.io/driver/postgres v1.5.9
