+++ gorilla-minitwit/src/internal/models/DeleteData.go
