+++ gorilla-minitwit/src/internal/models/RegisterData.go
